FCOChangeStuff 0.1.95 (updates since last GitHub version 0.1.87): 0.1.95 Added setting to position the cursor automatically into the lootwindow's 1st row if you loot something. Will only work if auto loot all is disabled!
0.1.94
Added settings to change the tooltips width and/or scale the text in the tooltips

0.1.93
Added: Setting to auto-decline a started group ready check. This is not declining if you are inside a dungeon/cyrodiil/IC or a raid. A keybind was added to toggle this setting.
Changed: LibAddonMenu-2.0 r32 needed, so please update! Custom control for LibShifterbox's "disabled sounds" will be created with new LAm possibilities now.
-Updated API and dependency versions

0.1.92
Added keybind to mute and restore (2nd press) the audio sfx volume.
This is a small helper for the "loud dragon sounds" in the trials. Muting and restoring the sfx volume seems to fix this problem temporarily until the next dragon shouts...

0.1.91
-Changed sound mute on mount to lower only the SFX volume not the total game volume
-You can choose via slider to what value the SFX sound should be lowered. To 0 is standard.

0.1.90
-Added setting to mute the game sound after mounting for x milliseconds (chosablke via a slider from 0 to 10 seconds) to supress these cries and flame and whatever mount related sounds

0.1.89
-Thie addon now needs LibShifterBox in addition!
-Updated API version
-Fixed MainMenu settings button added multiple times
-Added settings to disable chosen sounds. Scroll down to the bottom of the settings to see the list of sounds at the left. Move per drag&drop or click and use the > and < buttons to move the sounds from the left (avilable) to the right (disabled) sounds, or back.
Clicking the sound entries will play the sound. Not all soundss really play a sound so do not wonder if "nothing happens" upon clicking the entries!

0.1.88
-Added setting to show the equipped item + comparison panel at the bank deposit & withdraw
-Added setting to show the equipped item + comparison panel at the guild bank deposit & withdraw